[FlashRL 1/N] Add support for truncated importance sampling (#145)
# What does this PR do? 

Adds support for truncated importance sampling proposed in
https://fengyao.notion.site/flash-rl

To support this, we need to add a an optional `rollout_logprobs` fields
at the inference engine. Note that because the default path has no use
for rollout logprobs it is not obtained by default and is populated as
`None`. There is some ugliness with just being able to handle both these
cases (no logprobs and with logprobs) but it's fine for now.


TODO: 
- [x] Run E2E GPU tests for vllm and sglang
- [x] Test agent loop path with tis (requires token in, token out fix -
not taken up for now)

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>