get chat_end_index after env init (#184)
This PR moves the `chat_end_index` to after the `chat_history` has been
modified by the `env.init` function, to get the right number.

In our use case, we are injecting the system prompt in addition to the
user message in the `env.init`, so the length changes before and after.
In this case the `chat_end_index` is 1 instead of 2. We think it might
be causing some bugs down the line, leading to this assertion error:
```
AssertionError: Response ids and loss masks must have the same length, for sample 0 got 27 and 50
```

Resolves #183