[cleanup] Remove redundant sampling params code path (#234)
## What does this PR do?
Currently, the code base has two paths for plumbing the sampling params
in the training config to the inference engine requests. This PR reduces
it to one path for simplicity and to reduce surprises.

Path 1) At inference engine initialization time, e.g. in `main_base.py`,
the sampling params are passed into the constructor and saved as a
member variable of the engine class.

Path 2) Each call to `Generator.generate()` pass in the sampling params
from the `generator` config, which are then passed to the inference
engines as part of `InferenceEngineInput`.

This PR removes Path 1. Path 1 is less visible and more surprising.
Further, users are used to each inference request carrying sampling
params (rather than having a saved default in the engine itself), so it
is more natural.