[Fix][Generator] Use custom_chat_template in each step retokenization (#233)
Prior to this PR, in the re-tokenize chat history codepath, we did not
use `self.custom_chat_template` in some places (only used it in the
final tokenization). Though currently only Qwen3 enters this codepath,
which would not be affected, it is still good to apply the
`custom_chat_template` for consistency.