1/N Terminal Bench Integration (#239)
This PR is a WIP implementation of the Terminal Bench integration. We
are using the `sandboxes` repo instead of `terminal-bench` because
`sandboxes` is the new and improved version. The primary limitation of
the current integration is that we have hardcoded training to only use
one TBench task (`hello-world`). As `sandboxes` migrates more tasks from
TBench, we will add support for specifying the training set of tbench
tasks to use.

This PR is mostly a no-op for skyrl-train functionality because it
primarily adds/edits files in `/examples`.

---------

Co-authored-by: Charlie Ruan <charlieruan@berkeley.edu>
Co-authored-by: Sumanth R Hegde <39546518+SumanthRH@users.noreply.github.com>