[fix] Resolve CI error due to import error (#256)
We cannot import `VerifiersGenerator` unless `--with verifiers` is
included in the `uv` command. An error is thrown when pytest is
collecting all tests and tries to import `VerifiersGenerator`.

This PR moves the import of `VerifiersGenerator` in the test into the
test body so that it's only imported when executed.

Confirmed that Verifiers integration tests still pass with the commands
in the CI pipeline.