[Generator] Add trajectory ID and training phase information to GeneratorInput  (#293)
# What does this PR do?

Should close #288 

In the mini-swe-agent example we save all the trajectories (retained as
a list of messages) at each training step for easy viz. This is pretty
common. It's better to have this save logic here rather than in the
training loop because I want to do this before chat templating gets
applied.

However, currently we are just saving it based on instance id, and
overwriting the same file for different steps (and different
trajectories per instance). The main reason is that our current
Generator APIs don't allow for training step information / traj uids to
.generate.

To support saving trajectories in the Generator similar to the above
case, this PR does the following:
1. Add `trajectory_ids` to `GeneratorInput`: A trajectory ID is a unique
identifier for a trajectory, comprising the row id (UUID per row in the
training dataset) and a repetition id. This can also be easily extended
when we support re-using generated trajectories from older training
steps (add a step id to TrajectoryID). Note that `trajectory_ids` is not
a hard requirement for generator input - I might want to generate in the
context of running evaluation and I might not care about this - so it is
optional.
2. Rename `trajectory_id` in `InferenceEngineInput` to `session_id` to
avoid confusion
3. Add `batch_metadata` to `GeneratorInput`: To attach a global step and
other metadata to the batch, it is best to add this information as a
part of the GeneratorInput itself. This should also translate over for
more complex use-cases like asynchronous training where the generator
and the trainer can be operating at different steps.
4. Create a true UID per row: Previously, we generated UIDs for a batch
in `_prepare_generator_input`, mainly for purposes of GRPO. But now,
given that we want to makea TrajectoryID with a row ID and a repetition
ID, we need a actual unique ID per row. The previous approach will
createa new UID for a given row in each epoch, which is bad. So now we
generate a UID in `__getitem__` in `PromptDataset` which is simply
`str(row_idx)`

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>